Feature/backend user controller

See merge request s05-final/S05P31A402!24

+++ backend/src/main/java/com/ssacretary/db/entity/Count.java
@@ -15,7 +15,7 @@ public class Count {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "count_id")
-    private Long countId;
+    private int countId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "log_id")
